+++ src/interfaces/i-packages.ts
@@ -1,11 +1,26 @@
 /**
  * calamares module packages
  */
+interface operations {
+  packages: string[]
+}
 
 export interface IPackages {
   backend: string
-  operations: {
-    remove: string[]
-    try_install: string[]
-  }
+  operations: operations[]
 }
+
+/**
+# Debian/Buster et others
+# packages
+---
+backend: apt
+
+operations:
+- remove:
+  - calamares
+  - eggs
+- try_install:
+  - firefox-esr-$LOCALE
+
+ */