+++ src/krill/modules/packages.ts
@@ -27,48 +27,55 @@ export default async function packages(this: Sequence): Promise<void> {
         modulePath = '/etc/calamares/'
     }
     const config_file = `${this.installTarget}${modulePath}modules/packages.conf`
-
+    let remove: string[] = []
+    let install: string[] = []
+    
     if (fs.existsSync(config_file)) {
-        const pkgs = yaml.load(fs.readFileSync(config_file, 'utf-8')) as IPackages
-        if (pkgs.operations !== undefined) {
-            if (pkgs.backend === 'apt') {
+        const packages = yaml.load(fs.readFileSync(config_file, 'utf-8')) as IPackages
+        let operations = JSON.parse(JSON.stringify(packages.operations))
+        remove = operations[0].remove
+        if (operations.length > 1) {
+            install = operations[1].install
+        }
+        
+        if (operations !== undefined) {
+            if (packages.backend === 'apt') {
                 /**
                  * apt: Debian/Devuan/Ubuntu
                  */
-                if (pkgs.operations.remove !== undefined) {
+
+                if (remove.length > 0) {
                     let cmd = `chroot ${this.installTarget} apt-get purge -y `
-                    for (const elem of pkgs.operations.remove) {
+                    for (const elem of remove) {
                         cmd += elem + ' '
                     }
                     await exec(`${cmd} ${this.toNull}`, this.echo)
-                    Utils.pressKeyToExit(cmd)
                 }
 
-                if (pkgs.operations.try_install !== undefined) {
+                if (install.length > 0) {
                     let cmd = `chroot ${this.installTarget} apt-get install -y `
-                    for (const elem of pkgs.operations.try_install) {
+                    for (const elem of install) {
                         cmd += elem + ' '
                     }
                     await exec(`chroot ${this.installTarget} apt-get update ${this.toNull}`, this.echo)
                     await exec(`${cmd} ${this.toNull}`, this.echo)
-                    Utils.pressKeyToExit(cmd)
                 }
 
 
-            } else if (pkgs.backend === 'pacman') {
+            } else if (packages.backend === 'pacman') {
                 /**
                  * pacman: arch/manjaro
                  */
-                if (pkgs.operations.remove !== undefined) {
+                if (remove.length > 0) {
                     let cmd = `chroot ${this.installTarget} pacman -R\n`
-                    for (const elem of pkgs.operations.remove) {
+                    for (const elem of remove) {
                         cmd += elem + ' '
                     }
                     await exec(`${cmd} ${echoYes}`, this.echo)
                 }
 
-                if (pkgs.operations.try_install !== undefined) {
-                    for (const elem of pkgs.operations.try_install) {
+                if (install.length > 0) {
+                    for (const elem of install) {
                         await exec(`chroot ${this.installTarget} pacman -S ${elem}`, echoYes)
                     }
                 }
@@ -77,3 +84,29 @@ export default async function packages(this: Sequence): Promise<void> {
     }
 }
 
+
+/*
+const packages = yaml.load(fs.readFileSync(config_file, 'utf-8')) as IPackages
+let operations = JSON.parse(JSON.stringify(packages.operations))
+let remove: string[] = []
+let install: string[] = []
+
+remove = operations[0].remove
+if (operations.length > 1) {
+    install = operations[1].install
+}
+
+if (operations !== undefined) {
+    if (packages.backend === 'apt') {
+
+        if (remove.length > 0) {
+            let ctr = `chroot apt-get purge -y `
+            for (const elem of remove) {
+                ctr += elem + ' '
+            }
+            console.log(`${ctr}`)
+            console.log(`apt-get autoremove -y `)
+        }
+    }
+}
+*/